requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
PyJWT==2.10.1
soupsieve==2.7
typing_extensions==4.14.0
Werkzeug==3.1.3
